Organização dos componentes do projeto.
