+++ src/components/AppBar/app-bar.scss
@@ -1,0 +1,60 @@
+@import "../../styles/variables.scss";
+@import "../../styles/mixin.scss";
+
+.app-bar {
+  z-index: 999;
+  font-size: 14pt;
+  font-weight: bold;
+  color: white;
+  background-color: $color-primary;
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  padding: 0;
+  margin: 0;
+
+  &__container {
+    height: 70px;
+    display: flex;
+    align-items: center;
+    margin: 0 auto;
+    padding: 0 $app-padding;
+    max-width: $container-max-width;
+  }
+
+  &__brand {
+    flex: 1;
+    padding-left: 8px;
+  }
+
+  &__action {
+    color: white;
+    background-color: transparent;
+    border: none;
+    border-radius: 2px;
+    cursor: pointer;
+
+    &:focus {
+      outline: none;
+    }
+
+    &:hover {
+      background-color: $color-hover;
+    }
+
+    &--rotation {
+      @include rotation();
+    }
+
+    &--danger {
+      color: $color-danger;
+    }
+
+    & i {
+      display: block;
+    }
+  }
+
+  @include box-shadow(2);
+}
