+++ src/components/NavigationDrawer/navigation-drawer.scss
@@ -1,0 +1,57 @@
+@import "../../styles/mixin.scss";
+
+.navigation-drawer {
+  z-index: 9999;
+  background-color: white;
+  position: fixed;
+  top: 0;
+  left: -280px;
+  height: 100vh;
+  width: 260px;
+  transition: left 0.5s ease-in-out;
+
+  @include box-shadow(5);
+
+  &--open {
+    left: 0;
+  }
+
+  &__head {
+    height: 46px;
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+    padding-right: 12px;
+
+    &__button {
+      background-color: transparent;
+      border: none;
+      cursor: pointer;
+
+      & i {
+        display: block;
+      }
+    }
+  }
+
+  &__menu {
+    margin-top: 16px;
+    padding: 0 16px 0 8px;
+
+    &__item {
+      background-color: transparent;
+      border: none;
+      display: flex;
+      align-items: center;
+      vertical-align: middle;
+      width: 100%;
+      margin-bottom: 12px;
+      cursor: pointer;
+
+      & i {
+        display: block;
+        padding-right: 8px;
+      }
+    }
+  }
+}
