+++ src/components/Student/StudentList.js
@@ -2,6 +2,8 @@ import React from "react";
 
 import Student from "./Student";
 
+import "./student.scss";
+
 const StudentList = ({ students, onEdit, onDelete, onMove }) => (
   <div className="student-list">
     {students.map((student, index) => (
