+++ src/components/TheoryClass/NewTheoryClass.js
@@ -1,12 +1,14 @@
 import React from "react";
 
-export default class NewStudent extends React.Component {
+import "./theory-class.scss";
+
+export default class NewTheoryClass extends React.Component {
   state = {
     name: ""
   };
 
-  handleAdd(student) {
-    this.props.onAdd(student);
+  handleAdd(theoryClass) {
+    this.props.onAdd(theoryClass);
     this.setState({ name: "" });
   }
 
@@ -14,11 +16,11 @@ export default class NewStudent extends React.Component {
     const { name } = this.state;
 
     return (
-      <div className="new-student">
+      <div className="new-theory-class">
         <input
           type="text"
-          className="new-student__input"
-          placeholder="Digite o nome do novo aluno..."
+          className="new-theory-class__input"
+          placeholder="Digite o nome da nova turma..."
           value={name}
           onChange={event => this.setState({ name: event.target.value })}
           onKeyPress={event => {
@@ -26,9 +28,11 @@ export default class NewStudent extends React.Component {
           }}
         />
         <button
-          className="new-student__button"
+          className="new-theory-class__button"
           onClick={() =>
-            this.handleAdd(document.querySelector(".new-student__input").value)
+            this.handleAdd(
+              document.querySelector(".new-theory-class__input").value
+            )
           }
         >
           <i className="material-icons">save</i>
