+++ src/components/index.js
@@ -1,0 +1,13 @@
+import AppBar from "./AppBar/AppBar";
+import NavigationDrawer from "./NavigationDrawer/NavigationDrawer";
+import Error from "./Error/Error";
+import Enrollment from "./Enrollment/Enrollment";
+import EnrollmentList from "./Enrollment/EnrollmentList";
+import NewStudent from "./Student/NewStudent";
+import Student from "./Student/Student";
+import StudentList from "./Student/StudentList";
+import NewTheoryClass from "./TheoryClass/NewTheoryClass";
+import TheoryClass from "./TheoryClass/TheoryClass";
+import TheoryClassList from "./TheoryClass/TheoryClassList";
+
+export { AppBar, NavigationDrawer, Error, Enrollment, EnrollmentList, NewStudent, Student, StudentList, NewTheoryClass, TheoryClass, TheoryClassList };
