+++ src/containers/About.js
@@ -3,9 +3,9 @@ import React from "react";
 const About = () => (
   <div className="about">
     <h1>Autoescola Senna</h1>
-    <h2>Desafio Prático - Parte 1</h2>
+    <h2>Desafio Prático - Parte 2</h2>
     <p>
-      Implementação dos requisitos da aplicação definidos na Seção 3, Aula 27.
+      Implementação dos requisitos da aplicação definidos na Seção 4, Aula 44.
     </p>
   </div>
 );
