+++ src/index.scss
@@ -1,16 +1,6 @@
-@import "./mixin";
+@import "styles/mixin.scss";
+@import "styles/variables.scss";
 
-$app-padding: 20px;
-$container-max-width: 600px;
-$color-primary: darkslateblue;
-$color-danger: red;
-$color-success: green;
-$color-hover: rgba(
-  $color: black,
-  $alpha: 0.15
-);
-$icon-height: 24px;
-$icon-width: 24px;
 
 body {
   color: #212121;
@@ -45,63 +35,6 @@ button {
   }
 }
 
-.app-bar {
-  z-index: 999;
-  font-size: 14pt;
-  font-weight: bold;
-  color: white;
-  background-color: $color-primary;
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  padding: 0;
-  margin: 0;
-
-  &__container {
-    height: 70px;
-    display: flex;
-    align-items: center;
-    margin: 0 auto;
-    padding: 0 $app-padding;
-    max-width: $container-max-width;
-  }
-
-  &__brand {
-    flex: 1;
-    padding-left: 8px;
-  }
-
-  &__action {
-    color: white;
-    background-color: transparent;
-    border: none;
-    border-radius: 2px;
-    cursor: pointer;
-
-    &:focus {
-      outline: none;
-    }
-
-    &:hover {
-      background-color: $color-hover;
-    }
-
-    &--rotation {
-      @include rotation();
-    }
-
-    &--danger {
-      color: $color-danger;
-    }
-
-    & i {
-      display: block;
-    }
-  }
-
-  @include box-shadow(2);
-}
 
 .container {
   margin: 0 auto;
@@ -109,362 +42,16 @@ button {
   max-width: $container-max-width;
 }
 
-.new-student {
-  max-width: $container-max-width;
-  display: flex;
-  background-color: white;
-  padding: 16px 12px;
-  border-radius: 2px;
 
-  @include box-shadow(1);
 
-  &__input {
-    flex: 1;
-    border: none;
 
-    &:focus {
-      outline: none;
-    }
-  }
-
-  &__button {
-    background-color: transparent;
-    border: none;
-    padding: 0;
-    margin: 0;
-    height: $icon-height;
-    width: $icon-width;
-
-    &:hover {
-      background-color: $color-hover;
-    }
-
-    & i {
-      display: block;
-    }
-  }
-}
-
-.students {
-  &__container {
-    margin-bottom: 40px;
-  }
-
-  &__button {
-    background-color: whitesmoke;
-    border: none;
-    padding: 0;
-    margin: 0;
-    height: $icon-height;
-    width: $icon-width;
-
-    &__background {
-      background-color: whitesmoke;
-      border-radius: 45px;
-      width: 50px;
-      height: 50px;
-      position: relative;
-      float: right;
-      top: 24px;
-      right: 20px;
-    }
-
-    &:hover {
-      background-color: $color-hover;
-    }
-
-    &--floating {
-      font-size: 20px;
-      border-radius: 45px;
-      width: 50px;
-      height: 50px;
-
-      @include box-shadow(2);
-
-      &:focus {
-        outline: none;
-      }
-    }
-
-    & i {
-      display: block;
-    }
-  }
-}
-
-.student-list {
-  margin: 40px 0;
-}
-
-.student {
-  display: flex;
-  align-items: center;
-  margin-bottom: 20px;
-  background-color: white;
-  padding: 0px 12px;
-  border-radius: 2px;
-
-  @include box-shadow(1);
-
-  &__text {
-    flex: 1;
-  }
-
-  &__input {
-    flex: 1;
-    border-top: 0;
-    border-right: 0;
-    border-left: 0;
-    border-bottom: 1px solid silver;
-  }
-
-  &__button {
-    background-color: transparent;
-    border: none;
-    padding: 0;
-    margin: 0;
-    height: $icon-height;
-    width: $icon-width;
-
-    &:hover {
-      background-color: $color-hover;
-    }
-
-    &--cancel {
-      color: $color-danger;
-    }
-
-    &--success {
-      color: $color-success;
-    }
-
-    &--hidden {
-      visibility: hidden;
-    }
-
-    &--arrow {
-      padding: 0;
-      display: inline-block;
-      width: 100%;
-
-      &--none {
-        width: 1px;
-      }
-    }
-  }
-
-  &__arrows {
-    width: 24px;
-    height: 58px;
-    padding: 2px;
-
-    &--hidden {
-      visibility: hidden;
-      width: 1px;
-    }
-  }
-}
-
-.new-theory-class {
-  @extend .new-student;
-
-  &__input {
-    @extend .new-student__input;
-  }
-
-  &__button {
-    @extend .new-student__button;
-  }
-}
-
-.theory-class {
-  @extend .student;
-
-  &__input {
-    @extend .student__input;
-  }
-
-  &__text {
-    @extend .student__text;
-  }
-
-  &__button {
-    @extend .student__button;
-
-    &--cancel {
-      @extend .student__button--cancel;
-    }
-
-    &--success {
-      @extend .student__button--success;
-    }
-
-    &--arrow {
-      @extend .student__button--arrow;
-    }
-
-    &--hidden {
-      @extend .student__button--hidden;
-    }
-  }
-
-  &__arrows {
-    @extend .student__arrows;
-
-    &--hidden {
-      @extend .student__arrows--hidden;
-    }
-  }
-}
-
-.theory-classes {
-  @extend .students;
-
-  &__container {
-    @extend .students__container;
-  }
-
-  &__button {
-    @extend .students__button;
-
-    &__background {
-      @extend .students__button__background;
-    }
-
-    &--floating {
-      @extend .students__button--floating;
-    }
-  }
-}
-
-.theory-class-list {
-  @extend .student-list;
-}
-
-.enrollment {
-  @extend .student;
-
-  &__text {
-    padding: 16px 0px;
-
-    &--done {
-      @extend .student__text;
-      color: $color-success;
-    }
-
-    &--hidden {
-      visibility: hidden;
-    }
-  }
-
-  &__button {
-    @extend .student__button;
-
-    &--hidden {
-      display: none;
-    }
-  }
-}
-
-.enrollments {
-  @extend .students;
-
-  &__container {
-    @extend .students__container;
-  }
-
-  &__button {
-    @extend .students__button;
-
-    &__background {
-      @extend .students__button__background;
-    }
-  }
-}
-
-.enrollment-list {
-  @extend .student-list;
-}
 
 .about {
   text-align: center;
 }
 
-.error {
-  text-align: center;
 
-  &__button {
-    color: white;
-    background-color: transparent;
-    border: none;
-    border-radius: inherit;
-    padding: 10px 12px;
-    cursor: pointer;
 
-    &__background {
-      background-color: $color-primary;
-      width: fit-content;
-      border-radius: 2px;
-      display: inline-block;
-    }
-  }
-}
-
-.navigation-drawer {
-  z-index: 9999;
-  background-color: white;
-  position: fixed;
-  top: 0;
-  left: -280px;
-  height: 100vh;
-  width: 260px;
-  transition: left 0.5s ease-in-out;
-
-  @include box-shadow(5);
-
-  &--open {
-    left: 0;
-  }
-
-  &__head {
-    height: 46px;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    padding-right: 12px;
-
-    &__button {
-      background-color: transparent;
-      border: none;
-      cursor: pointer;
-
-      & i {
-        display: block;
-      }
-    }
-  }
-
-  &__menu {
-    margin-top: 16px;
-    padding: 0 16px 0 8px;
-
-    &__item {
-      background-color: transparent;
-      border: none;
-      display: flex;
-      align-items: center;
-      vertical-align: middle;
-      width: 100%;
-      margin-bottom: 12px;
-      cursor: pointer;
-
-      & i {
-        display: block;
-        padding-right: 8px;
-      }
-    }
-  }
-}
 
 .home {
   & i {
