+++ src/styles/variables.scss
@@ -1,0 +1,11 @@
+$app-padding: 20px;
+$container-max-width: 600px;
+$color-primary: darkslateblue;
+$color-danger: red;
+$color-success: green;
+$color-hover: rgba(
+  $color: black,
+  $alpha: 0.15
+);
+$icon-height: 24px;
+$icon-width: 24px;
